% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_header}
\alias{add_header}
\title{add_header
Generic helper function to make the first column the header.
This is linked to preparing the query data when manually entered in the app.}
\usage{
add_header(df)
}
\arguments{
\item{df}{user data.frame generated by manual data entry in the app.}
}
\value{
A re-formatted data.frame with header based on column names
}
\description{
add_header
Generic helper function to make the first column the header.
This is linked to preparing the query data when manually entered in the app.
}
\examples{
\dontrun{
add_header(user_df))
}
}
