% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_query_alleles_data.R
\name{add_query_alleles_data}
\alias{add_query_alleles_data}
\title{add_query_alleles_data
Adds the total number of alleles for the query and return a data.frame in the format
which will be viewed in the app - e.g.: allele values being comma separated}
\usage{
add_query_alleles_data(query, include_amel = TRUE)
}
\arguments{
\item{query}{query data}

\item{include_amel}{logical indicating whether to include Amelogenin in score computation}
}
\value{
A new query data.frame containing the number of total alleles in the data.frame format
needed by the app
}
\description{
add_query_alleles_data
Adds the total number of alleles for the query and return a data.frame in the format
which will be viewed in the app - e.g.: allele values being comma separated
}
\examples{
\dontrun{
add_query_alleles_data(query)
}
}
