% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scores.R
\name{add_scores}
\alias{add_scores}
\title{add_scores
Adds matching scores between query and reference to the reference data containing all markers}
\usage{
add_scores(
  ref_alleles_data,
  query_alleles_data,
  scoring_algorithm = c("tanabe", "masters"),
  masters_denominator = c("query", "reference")
)
}
\arguments{
\item{ref_alleles_data}{processed reference data - should be the output of \code{add_share_ref_alleles_data()}}

\item{query_alleles_data}{processed query data - should be the output of \code{add_query_alleles_data()}}

\item{scoring_algorithm}{choice of scoring algorithm: \code{tanabe} or \code{masters}}

\item{masters_denominator}{if \code{masters} is chosen in \code{scoring_algorithm}, choose the which denominator
to use in computation. Should be either \code{query} (original Masters) or \code{reference} (modified Masters)}
}
\value{
A new reference data.frame containing the matching scores between query and each marker in the reference.
}
\description{
add_scores
Adds matching scores between query and reference to the reference data containing all markers
}
\examples{
\dontrun{
add_scores(ref_alleles_data = add_share_ref_alleles_data(query),
query_alleles_data = add_query_alleles_data(query),
scoring_algorithm = "tanabe")
}
}
