% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_share_ref_alleles_data.R
\name{add_share_ref_alleles_data}
\alias{add_share_ref_alleles_data}
\title{add_share_ref_alleles_data
Adds the number of total alleles per marker and the number of shared alleles
between markers in the reference vs query. The output data.frame is also
transformed to the format which will be viewed in the app -
e.g.: allele values being comma separated}
\usage{
add_share_ref_alleles_data(
  query,
  include_amel = TRUE,
  reference = "app_reference"
)
}
\arguments{
\item{query}{query data}

\item{include_amel}{logical indicating whether to include Amelogenin in score computation}

\item{reference}{to be implemented in STR search. May be the built-in reference from the app
or a custom tumor line reference with the same STR markers used in the app reference}
}
\value{
A new reference data.frame containing the number of shared alleles (vs query)
and total alleles per marker in the data.frame format needed by the app
}
\description{
add_share_ref_alleles_data
Adds the number of total alleles per marker and the number of shared alleles
between markers in the reference vs query. The output data.frame is also
transformed to the format which will be viewed in the app -
e.g.: allele values being comma separated
}
\examples{
\dontrun{
add_share_ref_alleles_data(query)
}
}
