% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GbmPdxSTR_app.R
\name{GbmPdxSTR}
\alias{GbmPdxSTR}
\title{GbmPdxSTR
Runs the GbmPdxSTR application}
\usage{
GbmPdxSTR(...)
}
\value{
The shiny app instance
}
\description{
GbmPdxSTR
Runs the GbmPdxSTR application
}
\examples{
GbmPdxSTR()
}
