% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masters.R
\name{masters}
\alias{masters}
\title{masters
Masters match algorithm for STR profiling. Computes and return the score.
The denominator (total_alleles)}
\usage{
masters(shared_allele, total_allele)
}
\arguments{
\item{shared_allele}{The number of shared alleles between query and reference}

\item{total_allele}{The total alleles present in either the query (original Masters)
or the reference (modified Masters)}
}
\value{
Masters score
}
\description{
masters
Masters match algorithm for STR profiling. Computes and return the score.
The denominator (total_alleles)
}
\examples{
masters(shared_allele = 13, total_allele = 15)
}
