% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_multi_query.R
\name{process_multi_query}
\alias{process_multi_query}
\title{process_multi_query
Process multi-queries from a csv file}
\usage{
process_multi_query(input_path, ...)
}
\arguments{
\item{input_path}{the path to the multi-query csv file expected by the app}

\item{...}{params to be passed on to \code{process_query}}
}
\value{
A list of processed queries
}
\description{
process_multi_query
Process multi-queries from a csv file
}
\examples{
\dontrun{
process_multi_query("./data/multi_query_example.csv") #default params
process_multi_query("./data/multi_query_example.csv", include_amel = TRUE, scoring_algorithm = "tanabe") #same as default
process_multi_query("./data/multi_query_example.csv", include_amel = FALSE, scoring_algorithm = "masters", masters_denominator = "query")
}
}
