% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{process_query
Process single-query entries to produce a summarized data.frame containing STR matching scores}
\usage{
process_query(query, include_amel = TRUE, reference = "app_reference", ...)
}
\arguments{
\item{query}{query data}

\item{include_amel}{logical indicating whether to include Amelogenin in score computation}

\item{reference}{choice to use string \code{app_reference} to use the tumor reference included in the app or path to csv file containing custom reference}

\item{...}{params to be passed on to \code{add_scores} - typically linked to choice to scoring algorithm}
}
\value{
processed data frame containing STR markers from query and reference along with scores
}
\description{
process_query
Process single-query entries to produce a summarized data.frame containing STR matching scores
}
\examples{
\dontrun{
process_query(query, scoring_algorithm = "tanabe")
process_query(query, scoring_algorithm = "masters", masters_denominator = "query")
process_query(query, scoring_algorithm = "masters", masters_denominator = "reference")
}
}
