% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{process_upload}
\alias{process_upload}
\title{process_upload
Generic helper function to pre-process csv uploads to correctly format table required by the app.
It will also ensure that homozygous alleles are counted as one, and handles any STR aliases.}
\usage{
process_upload(input_path)
}
\arguments{
\item{input_path}{the path to the query csv file expected by the app}
}
\value{
A data.frame in the format that expected for further processing within the app
}
\description{
process_upload
Generic helper function to pre-process csv uploads to correctly format table required by the app.
It will also ensure that homozygous alleles are counted as one, and handles any STR aliases.
}
\examples{
\dontrun{
process_upload("./data/single_query_example.csv")
}
}
