% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_multi_query_workbook.R
\name{save_multi_query_workbook}
\alias{save_multi_query_workbook}
\title{save_multi_query_workbook
Saves a processed multi-query to an Excel spreadsheet where each resulting query is in a tab.}
\usage{
save_multi_query_workbook(
  multi_query_list,
  file_output_name = "GBM_PDX_STR_multi_query.xlsx"
)
}
\arguments{
\item{multi_query_list}{a list of processed queries - typically the output of \code{process_multi_query}}

\item{file_output_name}{output \code{xlsx} file name}
}
\value{
An excel spreadsheet which contains the multi-query results - one query per tab
}
\description{
save_multi_query_workbook
Saves a processed multi-query to an Excel spreadsheet where each resulting query is in a tab.
}
\examples{
\dontrun{
save_multi_query_workbook(process_multi_query("./data/multi_query_example.csv"))
# note nested `process_multi_query`
}
}
