% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_allele_stats.R
\name{summarize_allele_stats}
\alias{summarize_allele_stats}
\title{summarize_allele_stats
This function processes a list of alleles into a summarized data.frame (e.g.: sums alleles across marker etc.)
It's meant to be a helper function to other primary functions in the app}
\usage{
summarize_allele_stats(allele_input, new_column_name, perform_binding = TRUE)
}
\arguments{
\item{allele_input}{input data containing the total or shared number of alleles}

\item{new_column_name}{column name in output data.frame}

\item{perform_binding}{options to perform binding in input data - needed for reference data; optional for query}
}
\value{
A data frame of summarized allelic information
}
\description{
summarize_allele_stats
This function processes a list of alleles into a summarized data.frame (e.g.: sums alleles across marker etc.)
It's meant to be a helper function to other primary functions in the app
}
\examples{
\dontrun{
summarize_allele_stats(shared_alleles_list, new_column_name="shared_alleles")
}
}
