% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_multi_query.R
\name{summarize_multi_query}
\alias{summarize_multi_query}
\title{summarize_multi_query
Provides a summary of the scores from a multi-query search}
\usage{
summarize_multi_query(multi_query_list)
}
\arguments{
\item{multi_query_list}{a list of processed queries - typically the output of \code{process_multi_query}}
}
\value{
A data.frame which contains summary scores of each STR marker per sample.
}
\description{
summarize_multi_query
Provides a summary of the scores from a multi-query search
}
\examples{
\dontrun{
summarize_multi_query(process_multi_query("./data/multi_query_example.csv"))
# note nested `process_multi_query`
}
}
