% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanabe.R
\name{tanabe}
\alias{tanabe}
\title{tanabe
Tanabe match algorithm for STR profiling. Computes and return the score}
\usage{
tanabe(query_total_allele, reference_total_allele, shared_allele)
}
\arguments{
\item{query_total_allele}{The total alleles present in the query}

\item{reference_total_allele}{The total alleles present in the reference}

\item{shared_allele}{The number of shared alleles between query and reference}
}
\value{
Tanabe score
}
\description{
tanabe
Tanabe match algorithm for STR profiling. Computes and return the score
}
\examples{
tanabe(query_total_allele = 15, reference_total_allele = 16, shared_allele = 13)
}
