% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_marker_query.R
\name{validate_marker_query}
\alias{validate_marker_query}
\title{validate_marker_query
This function validates STR data entries and converts homozygous alleles to be counted as one for manual data entries}
\usage{
validate_marker_query(query, input_type = c("manual_entry", "csv"))
}
\arguments{
\item{query}{query data}

\item{input_type}{the data input type - either \code{manual_entry} for data input in the app or \code{csv} for file uploads}
}
\value{
the validated query
}
\description{
validate_marker_query
This function validates STR data entries and converts homozygous alleles to be counted as one for manual data entries
}
\examples{
\dontrun{
validate_marker_query(query, input_type = "manual_entry")
}
}
